requirements.txt
beautifulsoup4==4.10.0
jupyterlab==3.1.18
keras==2.6.0
matplotlib==3.4.3
Pillow==8.3.2
requests==2.26.0
tensorflow==2.6.0
tqdm==4.62.3